requirements.txt
pandas
dask
monkdb
shapely
rasterio
pyarrow
pyproj
gradio
llama-cpp-python
python-dotenv
mcp-monkdb